Consolidate CellPopConfig to legend
