+++ src/CellPopComponent.tsx
@@ -2,7 +2,6 @@ import React from "react";
 
 import Skeleton from "@mui/material/Skeleton";
 import { CellPopData, CellPopTheme } from "./cellpop-schema";
-import CellPopConfig from "./CellPopConfig";
 import { AxisConfig } from "./contexts/AxisConfigContext";
 import { Dimensions } from "./contexts/DimensionsContext";
 import { Providers } from "./contexts/Providers";
@@ -36,7 +35,6 @@ export const CellPop = ({
         xAxisConfig={xAxisConfig}
         yAxisConfig={yAxisConfig}
       >
-        <CellPopConfig />
         <VizContainer />
       </Providers>
     </div>
