+++ src/visx-visualization/Legend.tsx
@@ -13,6 +13,14 @@ import MenuItem from "@mui/material/MenuItem";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
 import { useEventCallback } from "@mui/material/utils";
 
+import Button from "@mui/material/Button";
+import ToggleButton from "@mui/material/ToggleButton";
+import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
+
+import { useSetTheme } from "../contexts/CellPopThemeContext";
+import { useFraction } from "../contexts/FractionContext";
+import { useSelectedDimension } from "../contexts/SelectedDimensionContext";
+
 export default function Legend() {
   const {
     scale: colors,
@@ -29,6 +37,37 @@ export default function Legend() {
   const handleThemeChange = useEventCallback((e: SelectChangeEvent) => {
     setHeatmapTheme(e.target.value as HeatmapTheme);
   });
+
+  const { currentThemeName: currentTheme, setTheme } = useSetTheme();
+  const { fraction, setFraction } = useFraction();
+  const { selectedDimension, setSelectedDimension } = useSelectedDimension();
+
+  const undo = useEventCallback(() => {
+    console.warn("Not yet implemented");
+  });
+  const changeTheme = useEventCallback(
+    (_: unknown, newTheme: "dark" | "light") => {
+      if (newTheme) {
+        setTheme(newTheme);
+      }
+    },
+  );
+
+  const changeFraction = useEventCallback(
+    (_: unknown, newFraction: boolean) => {
+      if (newFraction !== null) {
+        setFraction(newFraction);
+      }
+    },
+  );
+
+  const changeSelectedDimension = useEventCallback(
+    (event: React.MouseEvent<HTMLElement>) => {
+      setSelectedDimension(
+        (event.target as HTMLInputElement).value as "X" | "Y",
+      );
+    },
+  );
 
   return (
     <Stack
@@ -73,6 +112,43 @@ export default function Legend() {
           ))}
         </Select>
       </FormControl>
+      <Stack direction="row" spacing={2}>
+        <Button variant="outlined" onClick={undo}>
+          Undo
+        </Button>
+        <ToggleButtonGroup
+          color="primary"
+          value={selectedDimension}
+          exclusive
+          onChange={changeSelectedDimension}
+          aria-label="Axis"
+        >
+          <ToggleButton value="X">X</ToggleButton>
+          <ToggleButton value="Y">Y</ToggleButton>
+        </ToggleButtonGroup>
+      </Stack>
+      <Stack direction="row" spacing={2}>
+        <ToggleButtonGroup
+          color="primary"
+          value={fraction}
+          exclusive
+          onChange={changeFraction}
+          aria-label="Fraction"
+        >
+          <ToggleButton value={false}>Count</ToggleButton>
+          <ToggleButton value={true}>Fraction</ToggleButton>
+        </ToggleButtonGroup>
+        <ToggleButtonGroup
+          color="primary"
+          value={currentTheme}
+          exclusive
+          onChange={changeTheme}
+          aria-label="Theme"
+        >
+          <ToggleButton value="light">Light</ToggleButton>
+          <ToggleButton value="dark">Dark</ToggleButton>
+        </ToggleButtonGroup>
+      </Stack>
     </Stack>
   );
 }
